+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed an issue with asset/binary files not working correctly with codegen templates.
+- Fixed an issue where `moon upgrade` would require a workspace.
+
 ## 1.13.0
 
 #### 🚀 Updates